Fix bug: deletion of study

+++ modules/common/app/daos/common/ComponentResultDao.java
@@ -36,6 +36,7 @@ public class ComponentResultDao extends AbstractDao {
     }
 
     public void removeAll(List<ComponentResult> componentResultList) {
+        if (componentResultList.isEmpty()) return;
         String queryStr = "DELETE FROM ComponentResult cr WHERE cr in :crList";
         Query query = jpa.em().createQuery(queryStr);
         query.setParameter("crList", componentResultList).executeUpdate();
